+++ SDK/osx/backendless/include/Types.h
@@ -17,6 +17,7 @@
     NSMutableDictionary	*abstractMappings;
     NSMutableDictionary	*clientMappings;
     NSMutableDictionary	*serverMappings;
+    NSMapTable<Class, NSMutableDictionary *> *propertyMappings;
 }
 
 @property (nonatomic, retain) NSString *swiftClassPrefix;
@@ -32,6 +33,8 @@
 -(NSString *)getClientClassForServerType:(NSString *)serverClassName;
 -(NSString *)objectMappedClassName:(id)obj;
 -(NSString *)typeMappedClassName:(Class)type;
+-(void)addClientPropertyMappingForClass:(Class)clientClass columnName:(NSString *)columnName propertyName:(NSString *)propertyName;
+-(NSDictionary *)getPropertiesMappingForClientClass:(Class)clientClass;
 // type reflection
 +(NSString *)objectClassName:(id)obj;
 +(NSString *)typeClassName:(Class)type;